src/test/java/baseball/domain/BaseballGameTest.java
package baseball.domain;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

@DisplayName("야구 게임 테스트")
public class BaseballGameTest {

    @Test
    @DisplayName("야구 게임 생성")
    void create() {
        // when
        BaseballGame baseballGame = BaseballGame.create();

        // then
        assertAll(
                () -> assertThat(baseballGame.getBaseballNumbers()).isNotNull(),
                () -> assertThat(baseballGame.getBaseballNumbers().getBaseballNumbers()[0]).isNotNull(),
                () -> assertThat(baseballGame.getBaseballNumbers().getBaseballNumbers()[1]).isNotNull(),
                () -> assertThat(baseballGame.getBaseballNumbers().getBaseballNumbers()[2]).isNotNull()
        );
    }
}
